% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_data.R
\name{glue_data}
\alias{glue_data}
\title{Glues a list of data.frames into one data.frame if the they have the same headers or column size as the first data.frame}
\usage{
glue_data(input_data, check_header_names = TRUE)
}
\arguments{
\item{input_data}{List data.frame}

\item{check_header_names}{logical. True to Check the files are the same by column names or False for number of columns}
}
\value{
data.frame
}
\description{
Glues a list of data.frames into one data.frame if the they have the same headers or column size as the first data.frame
}
\examples{

\donttest{

df_1 <- data.frame(a = c(1,2,3,4))
df_2 <- data.frame(a = c(5,6,7,8))
df_3 <- data.frame(b = c(9,10,11,12))

df_list <- list(df_1, df_2, df_3)

final_df_one <- glue_data(df_list)
# data.frame(a = c(1,2,3,4,5,6,7,8))

final_df_two <- glue_data(df_list, check_header_names = FALSE)
# data.frame(a = c(1,2,3,4,5,6,7,8,9,10,11,12))

}
}
