% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glue_files_from_folder.R
\name{glue_from_folder}
\alias{glue_from_folder}
\title{Collects files from folder and joins the data into one data frame}
\usage{
glue_from_folder(
  folder_path = "",
  pattern = "*.csv",
  check_header_names = TRUE,
  first_file = NA_character_,
  func = read.csv,
  ...
)
}
\arguments{
\item{folder_path}{character vector.}

\item{pattern}{character.  Regex pattern used in list.files.}

\item{check_header_names}{logical. True to Check the files are the same by column names or False for number of columns.}

\item{first_file}{character vector. Choose the file name that all the other files check for the same column names or number of columns.}

\item{func}{function. The function to use to open the files from the folder.  Default is csv.read.  User could use external packages to open files.}

\item{...}{Extra arguments. Can be added that will be passed into the func param.}
}
\value{
data.frame
}
\description{
Collects files from folder and joins the data into one data frame
}
\examples{

\donttest{

# df <- glue_from_folder("example_path/")

# df <- glue_from_folder(folder_path = "example_path/",
#                      pattern = "*example.csv",
#                        check_header_names = FALSE,
#                        first_file = "first_file.csv",
#                        func=read.csv,
#                        stringsAsFactors=FALSEglue
#                       )
}
}
