% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{indexOfDiffDataframeInVector}
\alias{indexOfDiffDataframeInVector}
\title{adds the index number of the data frames from a vector to remove in a int vector.}
\usage{
indexOfDiffDataframeInVector(
  indexOfDataframesToRemove,
  dataframeVector,
  checkHeaderNames
)
}
\arguments{
\item{indexOfDataframesToRemove}{an empty int vector passed by reference for index numbers to be added if required.}

\item{dataframeVector}{vector of data frames.}

\item{checkHeaderNames}{bool true if header names required checking for the helper function dataFrameHeadersCheck.}
}
\value{
void.
}
\description{
adds the index number of the data frames from a vector to remove in a int vector.
}
