% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{keepDataframesWithSameHeaders}
\alias{keepDataframesWithSameHeaders}
\title{removes data frames from a vector if they do not have the same header structure as the first data frame in the vector.}
\usage{
keepDataframesWithSameHeaders(dataframeVector, checkHeaderNames)
}
\arguments{
\item{dataframeVector}{vector of data frames.}

\item{checkHeaderNames}{bool true if header names required checking for the helper function dataFrameHeadersCheck.}
}
\value{
vector of data frames.
}
\description{
removes data frames from a vector if they do not have the same header structure as the first data frame in the vector.
}
