% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{moveToIndexOne}
\alias{moveToIndexOne}
\title{Moves a string in a charcater vector to the start of the vector if it matches the string input.}
\usage{
moveToIndexOne(x, chrVector)
}
\arguments{
\item{x}{String The string value to match in the character vector to move to the front.}

\item{chrVector}{character vector.}
}
\value{
character vector.
}
\description{
Moves a string in a charcater vector to the start of the vector if it matches the string input.
}
