% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{removeDifferentDataframes}
\alias{removeDifferentDataframes}
\title{removes data frames from a vector using the index positions from a vector of ints.
Helper function for keepDataframesWithSameHeaders.}
\usage{
removeDifferentDataframes(indexOfDataframesToRemove, dataframeVector)
}
\arguments{
\item{indexOfDataframesToRemove}{an int vector passed by reference with the position of the data frames in a vector to be removed.}

\item{dataframeVector}{vector of data frames.}
}
\value{
void.
}
\description{
removes data frames from a vector using the index positions from a vector of ints.
Helper function for keepDataframesWithSameHeaders.
}
