% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sameHeaders}
\alias{sameHeaders}
\title{checks the same headers are in both data frames past into the function.
Helper function for dataFrameHeadersCheck.}
\usage{
sameHeaders(leftDataFrame, rightDataFrame)
}
\arguments{
\item{leftDataFrame}{data frame.}

\item{rightDataFrame}{data frame.}
}
\value{
bool true if the data frames have the same headers else false.
}
\description{
checks the same headers are in both data frames past into the function.
Helper function for dataFrameHeadersCheck.
}
