% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setDataframeColNamesToFirstDataframe}
\alias{setDataframeColNamesToFirstDataframe}
\title{Sets all the data frames in a vector to have the same headers as the first data frame in the vector.}
\usage{
setDataframeColNamesToFirstDataframe(dataframeVector)
}
\arguments{
\item{dataframeVector}{vector of data frames.}
}
\value{
Vector of data frames with all the column names the same as the first data frame.
}
\description{
Sets all the data frames in a vector to have the same headers as the first data frame in the vector.
}
